% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aor_pairs.R
\name{aor_pairs}
\alias{aor_pairs}
\title{deliver the rxy values for the actual, optimal, and random pairs}
\usage{
aor_pairs(SP, Rxy)
}
\arguments{
\item{SP}{two column tibble with columns Female and Male giving the actual spawn pairs}

\item{Rxy}{the Rxy matrix in tidy format. (The output of \code{\link{read_kinship_matrix}}).}
}
\description{
deliver the rxy values for the actual, optimal, and random pairs
}
