% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sp_rxy_ids.R
\name{check_sp_rxy_ids}
\alias{check_sp_rxy_ids}
\title{check to make sure that all the spawn pair IDs occur in the rxys}
\usage{
check_sp_rxy_ids(SP, Rxy)
}
\arguments{
\item{SP}{spawn pairs tibble}

\item{Rxy}{rxy matrix}
}
\description{
Makes sure that every female and male spawn pair have an rxy.
I will add other checks here as they come up as problems.
}
