% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-computeRxy-output.R
\name{clean_computeRxy_output}
\alias{clean_computeRxy_output}
\title{Clean up ComputeRxy() output in order to use in aor_pairs}
\usage{
clean_computeRxy_output(D)
}
\arguments{
\item{D}{a tibble holding the output of \code{\link{ComputeRxy}()}. This should be
a tibble with columns named \code{ind1}, \code{ind2}, and \code{quellergt}}
}
\description{
ComputeRxy() returns an rxy value for every single pair in the data
set, but when we compare Actual, to Optimal, to Realized rxys
we want to only focus on the Males and Females denoted by a leading "F"
or "M" in their names, respectively.  This function filters things
down to just those individuals, and then also makes a column Female and
a Column Male and another called rxy.  This might be necessary for a number of
different things like creating spawning matrices.
}
