% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{computeRxy_example_output}
\alias{computeRxy_example_output}
\title{output from running computeRxy example}
\format{a tibble with columns ind1, ind2, and quellergt.  The
third column is the Rxy value.}
\usage{
computeRxy_example_output
}
\description{
This takes too long to run, but is needed for other examples.
}
\keyword{datasets}
