% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_kinship_matrix.R
\name{read_kinship_matrix}
\alias{read_kinship_matrix}
\title{read a tab-delimited kinship matrix into a tidy tibble of rxy values}
\usage{
read_kinship_matrix(path, skip = 6)
}
\arguments{
\item{path}{path to the file}

\item{skip}{number of lines to skip in the beginning
This forces the first column to be a female and the second a male.
So, our null distribution will be between males and females (not
between females and females, for exmample).}
}
\description{
It has to be the full symmetric matrix.  Can't be the lower half. Males have
to be named "MXXX" and females "FXXX" where XXX are some numbers or
characters or whatever.  The main thing is that females have to start
with an F and males with an M.
}
\examples{
kfile <- system.file("extdata/IGH_W1718_master_Kinsh_res.txt.gz", package = "CohoBroodstock")
}
