% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_column2kinship.R
\name{two_column2kinship}
\alias{two_column2kinship}
\title{read in a two-column formatted genotype file and turn it into something KINSHIP can read}
\usage{
two_column2kinship(TCF, outdir = NULL)
}
\arguments{
\item{TCF}{path to the two-column formatted file}

\item{outdir}{name of output directory to write the kinship file to.  If this is left unspecified
(or is passed in as NULL), then the output directory is the same as that of the input two-column
genotype file.}
}
\value{
Invisibly returns the data frame formatted for Kinship.  But, also writes the thing
out, so typically you won't use the return value.
}
\description{
Nothing very cool about this.  Just some formatting.  It will read a file in and then
create a new file in the same spot as the other one with "_kinship_input" appended to it.
}
\details{
Note that some of the files I have gotten from Libby have empty columns on the end (probably
a microsoft excel thing.)  So, I will drop any columns that start with X followed by a number.
}
\examples{
# get path to example file in package directory
tcffile <- system.file("extdata", "WSH_W1718_v5_two_column_data.txt", package = "CohoBroodstock")

# print it to see what that looks like:
tcffile

# now make a Kinship input version in the current directory
two_column2kinship(tcffile, outdir = ".")

# Note that when you do this you will get a few error messages:
# 1. missing column names filled in --- This is because the input file has two
# empty columns in it (i.e., it has the tabs but no content).
# 2. The two column names for each locus are the same and they get deduplicated.
# 3. See #1.

}
